refactor: improve module division
