+++ src/core/ffmpeg/check-install.test.ts
