+++ src/core/ffmpeg/check-install.ts
