+++ src/utils/date.ts
@@ -1,0 +1,22 @@
+export function getCurrentDateTime() {
+  const date = new Date();
+  const year = date.getFullYear();
+  const month = zeroPad(date.getMonth() + 1);
+  const day = zeroPad(date.getDate());
+  const hours = zeroPad(date.getHours());
+  const minutes = zeroPad(date.getMinutes());
+  const seconds = zeroPad(date.getSeconds());
+  return `${year}${month}${day}${hours}${minutes}${seconds}`;
+}
+
+export function formatSeconds(totalSeconds: number): string {
+  const hours = Math.floor(totalSeconds / 3600);
+  const remainingSeconds = totalSeconds % 3600;
+  const minutes = Math.floor(remainingSeconds / 60);
+  const seconds = remainingSeconds % 60;
+  return `${zeroPad(hours)}:${zeroPad(minutes)}:${zeroPad(seconds)}`;
+}
+
+function zeroPad(num: number) {
+  return num.toString().padStart(2, "0");
+}
