+++ src/utils/sanitize.ts
@@ -1,0 +1,5 @@
+import path from "node:path";
+
+export function sanitizePathLikeInput(rawInput: string) {
+  return rawInput.trim().replace(/[\\/]+/g, path.sep);
+}
